PR feedback

+++ src/base/reports/JsonReporter.hpp
@@ -24,30 +24,28 @@ public:
     void report(const LatencyReport& rep) override {
         printer_.write("{\"type\": \"latency\", \"device\": \"{}\", \"timestamp\": {},"
                        " \"sw_hw\": {}, \"hw\": {}, \"hw_avg\": {}}",
-            dev_name_, (unsigned long long)wallclock_timestamp_ns(), rep.sw_hw, rep.hw,
-            rep.hw_avg);
+            dev_name_, wallclock_timestamp_ns(), rep.sw_hw, rep.hw, rep.hw_avg);
     }
 
     void report(const LossReport& rep) override {
         printer_.write("{\"type\": \"losses\", \"device\": \"{}\", \"timestamp\": {},"
                        " \"rate\": {}, \"rate_avg\": {}, \"ratio\": {}}",
-            dev_name_, (unsigned long long)wallclock_timestamp_ns(), rep.loss_rate,
-            rep.loss_rate_avg, rep.loss_ratio);
+            dev_name_, wallclock_timestamp_ns(), rep.loss_rate, rep.loss_rate_avg,
+            rep.loss_ratio);
     }
 
     void report(const IOJitterReport& rep) override {
         printer_.write("{\"type\": \"io_jitter\", \"device\": \"{}\", \"timestamp\": {},"
                        " \"swdev_avg\": {}, \"swdev_per\": {},"
                        " \"hwdev_avg\": {}, \"hwdev_per\": {}}",
-            dev_name_, (unsigned long long)wallclock_timestamp_ns(), rep.swdev_avg,
-            rep.swdev_per, rep.hwdev_avg, rep.hwdev_per);
+            dev_name_, wallclock_timestamp_ns(), rep.swdev_avg, rep.swdev_per,
+            rep.hwdev_avg, rep.hwdev_per);
     }
 
     void report(const IODelayReport& rep) override {
         printer_.write("{\"type\": \"io_delay\", \"device\": \"{}\", \"timestamp\": {},"
                        " \"sw_avg\": {}, \"hw_avg\": {}}",
-            dev_name_, (unsigned long long)wallclock_timestamp_ns(), rep.sw_avg,
-            rep.hw_avg);
+            dev_name_, wallclock_timestamp_ns(), rep.sw_avg, rep.hw_avg);
     }
 
 private:
